tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.strokeStyle.colormix.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.strokeStyle.colormix</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.strokeStyle.colormix</h1>
<p class="desc">color-mix works as color input</p>


<script>
var t = async_test("color-mix works as color input");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.strokeStyle = "color-mix(in srgb, red, blue)";
  _assertSame(ctx.strokeStyle, 'color(srgb 0.5 0 0.5)', "ctx.strokeStyle", "'color(srgb 0.5 0 0.5)'");
  ctx.strokeStyle = "color-mix(in srgb, red, color(srgb 1 0 0))";
  _assertSame(ctx.strokeStyle, 'color(srgb 1 0 0)', "ctx.strokeStyle", "'color(srgb 1 0 0)'");
  t.done();

});
</script>
